requirements.txt
asgiref==3.4.1
Django==3.2.8
django-bootstrap==0.2.4
django-filter==21.1
django-mathfilters==1.0.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
